+++ desc/equilibrium/equilibrium.py
@@ -66,6 +66,7 @@ from .utils import (
     ensure_consistent_profile_eq_resolution,
     parse_axis,
     parse_profile,
+    parse_section,
     parse_surface,
 )
 
@@ -126,13 +127,16 @@ class Equilibrium(IOAble, Optimizable):
         Anisotropic pressure profile or array of mode numbers and spectral coefficients.
         Default is a PowerSeriesProfile with zero anisotropic pressure.
     surface: Surface or ndarray shape(k,5) (optional)
-        Fixed boundary surface shape, as a Surface object or array of
+        LCFS boundary shape, as a Surface object or array of
         spectral mode numbers and coefficients of the form [l, m, n, R, Z].
         Default is a FourierRZToroidalSurface with major radius 10 and minor radius 1
     axis : Curve or ndarray shape(k,3) (optional)
         Initial guess for the magnetic axis as a Curve object or ndarray
         of mode numbers and spectral coefficients of the form [n, R, Z].
         Default is the centroid of the surface.
+    xsection : ZernikeRZToroidalSection (optional)
+        Cross-section at zeta=0 as a ZernikeRZToroidalSection object. Can be supplied
+        only if the `surface` parameter is not given.
     sym : bool (optional)
         Whether to enforce stellarator symmetry. Default surface.sym or False.
     spectral_indexing : str (optional)
@@ -165,6 +169,7 @@ class Equilibrium(IOAble, Optimizable):
         "_L_basis",
         "_surface",
         "_axis",
+        "_xsection",
         "_pressure",
         "_iota",
         "_current",
@@ -174,7 +179,6 @@ class Equilibrium(IOAble, Optimizable):
         "_atomic_number",
         "_anisotropy",
         "_spectral_indexing",
-        "_bdry_mode",
         "_L_grid",
         "_M_grid",
         "_N_grid",
@@ -218,6 +222,7 @@ class Equilibrium(IOAble, Optimizable):
         anisotropy=None,
         surface=None,
         axis=None,
+        xsection=None,
         sym=None,
         spectral_indexing=None,
         check_orientation=True,
@@ -262,17 +267,34 @@ class Equilibrium(IOAble, Optimizable):
             ValueError,
             f"sym should be one of True, False, None, got {sym}",
         )
-        self._sym = bool(setdefault(sym, getattr(surface, "sym", False)))
+        errorif(
+            surface is not None and xsection is not None,
+            ValueError,
+            "Cannot specify both surface and xsection",
+        )
+        if surface is not None:
+            self._sym = bool(setdefault(sym, getattr(surface, "sym", False)))
+        elif xsection is not None:
+            self._sym = bool(setdefault(sym, getattr(xsection, "sym", False)))
+        else:
+            self._sym = bool(setdefault(sym, False))
         self._R_sym = "cos" if self.sym else False
         self._Z_sym = "sin" if self.sym else False
-
-        # surface
-        self._surface, self._bdry_mode = parse_surface(
-            surface, self.NFP, self.sym, self.spectral_indexing
-        )
-
-        # magnetic axis
-        self._axis = parse_axis(axis, self.NFP, self.sym, self.surface)
+        # Parse surface, magnetic axis and cross-section
+        if xsection is not None:
+            self._xsection = parse_section(
+                xsection, sym=self.sym, spectral_indexing=self.spectral_indexing
+            )
+            self._surface = parse_surface(surface, self.NFP, self.sym)
+            self._axis = parse_axis(axis, self.NFP, self.sym, xsection=self.xsection)
+        else:
+            self._surface = parse_surface(surface, self.NFP, self.sym)
+            self._axis = parse_axis(axis, self.NFP, self.sym, self.surface)
+            self._xsection = parse_section(
+                surface=self.surface,
+                sym=self.sym,
+                spectral_indexing=self.spectral_indexing,
+            )
 
         # resolution
         L = check_nonnegint(L, "L")
@@ -283,12 +305,16 @@ class Equilibrium(IOAble, Optimizable):
         N_grid = check_nonnegint(N_grid, "N_grid")
 
         self._N = int(setdefault(N, self.surface.N))
-        self._M = int(setdefault(M, self.surface.M))
+        self._M = (
+            int(setdefault(M, self.surface.M))
+            if surface is not None
+            else int(setdefault(M, self.xsection.M))
+        )
         self._L = int(
             setdefault(
                 L,
                 max(
-                    self.surface.L,
+                    self.xsection.L,
                     self.M if (self.spectral_indexing == "ansi") else 2 * self.M,
                 ),
             )
@@ -299,6 +325,7 @@ class Equilibrium(IOAble, Optimizable):
 
         self._surface.change_resolution(self.L, self.M, self.N, sym=self.sym)
         self._axis.change_resolution(self.N, sym=self.sym)
+        self._xsection.change_resolution(self.L, self.M)
 
         # bases
         self._R_basis = FourierZernikeBasis(
@@ -406,6 +433,11 @@ class Equilibrium(IOAble, Optimizable):
             ValueError,
             "Surface and Equilibrium must have the same symmetry",
         )
+        errorif(
+            self.sym != self.xsection.sym,
+            ValueError,
+            "Cross-section and Equilibrium must have the same symmetry",
+        )
         self._R_lmn = np.zeros(self.R_basis.num_modes)
         self._Z_lmn = np.zeros(self.Z_basis.num_modes)
         self._L_lmn = np.zeros(self.L_basis.num_modes)
@@ -416,7 +448,15 @@ class Equilibrium(IOAble, Optimizable):
             self.Z_lmn = kwargs.pop("Z_lmn")
             self.L_lmn = kwargs.pop("L_lmn", jnp.zeros(self.L_basis.num_modes))
         else:
-            self.set_initial_guess(ensure_nested=ensure_nested)
+            if xsection is None:
+                # For none Poincare BC problems, we need to set the xsection
+                # from the initial guess.
+                self.set_initial_guess(ensure_nested=ensure_nested, lcfs=True)
+                self._xsection = self.get_surface_at(zeta=0)
+            else:
+                self.set_initial_guess(ensure_nested=ensure_nested, lcfs=False)
+                self._surface = self.get_surface_at(rho=1.0)
+                self._axis = self.get_axis()
         if check_orientation:
             ensure_positive_jacobian(self)
         if kwargs.get("check_kwargs", True):
@@ -441,6 +481,11 @@ class Equilibrium(IOAble, Optimizable):
             # Need to rebuild derivative matrices to get higher order derivatives
             # on equilibrium's saved before GitHub pull request #586.
             self.current._transform = self.current._get_transform(self.current.grid)
+        if self._xsection is None:
+            # eq.xsection property was added with Poincare BC support
+            self._xsection = self.get_surface_at(zeta=0)
+        if not hasattr(self.surface, "_rho"):
+            self.surface._rho = 1.0
 
         # ensure things that should be ints are ints
         self._L = int(self._L)
@@ -470,6 +515,9 @@ class Equilibrium(IOAble, Optimizable):
             "Ti_l",
             "Zeff_l",
             "a_lmn",
+            "Rp_lmn",
+            "Zp_lmn",
+            "Lp_lmn",
             "Ra_n",
             "Za_n",
             "Rb_lmn",
@@ -478,7 +526,6 @@ class Equilibrium(IOAble, Optimizable):
             "G",
             "Phi_mn",
         )
-        assert sorted(args) == sorted(arg_order)
         return [arg for arg in arg_order if arg in args]
 
     def __repr__(self):
@@ -492,7 +539,7 @@ class Equilibrium(IOAble, Optimizable):
             )
         )
 
-    def set_initial_guess(self, *args, ensure_nested=True):
+    def set_initial_guess(self, *args, ensure_nested=True, lcfs=True):
         """Set the initial guess for the flux surfaces, eg R_lmn, Z_lmn, L_lmn.
 
         Parameters
@@ -514,6 +561,11 @@ class Equilibrium(IOAble, Optimizable):
             If True, and the default initial guess does not produce nested surfaces,
             run a small optimization problem to attempt to refine initial guess to
             improve coordinate mapping.
+        lcfs : bool
+            If True, the initial guess will be created by scaling down the LCFS
+            surface(rho=1.0), to form nested surfaces. If False, the initial guess
+            will be created by revolving the given Poincare section(zeta=0) resulting in
+            an axisymmetric equilibrium. Default is True.
 
         Examples
         --------
@@ -553,7 +605,7 @@ class Equilibrium(IOAble, Optimizable):
         >>> equil.set_initial_guess(nodes, R, Z, lambda)
 
         """
-        set_initial_guess(self, *args, ensure_nested=ensure_nested)
+        set_initial_guess(self, *args, ensure_nested=ensure_nested, lcfs=lcfs)
 
     def copy(self, deepcopy=True):
         """Return a (deep)copy of this equilibrium."""
@@ -644,6 +696,7 @@ class Equilibrium(IOAble, Optimizable):
             self.L, self.M, self.N, NFP=self.NFP, sym=self.sym
         )
         self.axis.change_resolution(self.N, NFP=self.NFP, sym=self.sym)
+        self.xsection.change_resolution(L=self.L, M=self.M, NFP=self.NFP, sym=self.sym)
 
         self._R_lmn = copy_coeffs(self.R_lmn, old_modes_R, self.R_basis.modes)
         self._Z_lmn = copy_coeffs(self.Z_lmn, old_modes_Z, self.Z_basis.modes)
@@ -726,12 +779,22 @@ class Equilibrium(IOAble, Optimizable):
             return surface
 
         if zeta is not None:
-            assert (zeta >= 0) and (zeta <= 2 * np.pi)
-            surface = ZernikeRZToroidalSection(sym=self.sym, zeta=zeta)
+            errorif(
+                not ((zeta >= 0) and (zeta <= 2 * np.pi / self.NFP)),
+                ValueError,
+                "zeta must be between 0 and {2 * np.pi / self.NFP} (2pi/NFP), got "
+                + f"{zeta}. Did you forgot to divide by NFP?",
+            )
+            # only zeta=0 and zeta=pi/NFP have stellarator symmetry
+            surface_sym = self.sym if zeta == 0 or zeta == np.pi / self.NFP else False
+            surface = ZernikeRZToroidalSection(
+                sym=surface_sym, zeta=zeta % (2 * np.pi / self.NFP)
+            )
             surface.change_resolution(self.L, self.M)
 
             AR = np.zeros((surface.R_basis.num_modes, self.R_basis.num_modes))
             AZ = np.zeros((surface.Z_basis.num_modes, self.Z_basis.num_modes))
+            AL = np.zeros((surface.L_basis.num_modes, self.L_basis.num_modes))
 
             for i, (l, m, n) in enumerate(self.R_basis.modes):
                 j = np.argwhere(
@@ -750,10 +813,22 @@ class Equilibrium(IOAble, Optimizable):
                     )
                 )
                 AZ[j, i] = fourier(zeta, n, self.NFP)
-            Rb = AR @ self.R_lmn
-            Zb = AZ @ self.Z_lmn
-            surface.R_lmn = Rb
-            surface.Z_lmn = Zb
+
+            for i, (l, m, n) in enumerate(self.L_basis.modes):
+                j = np.argwhere(
+                    np.logical_and(
+                        surface.L_basis.modes[:, 0] == l,
+                        surface.L_basis.modes[:, 1] == m,
+                    )
+                )
+                AL[j, i] = fourier(zeta, n, self.NFP)
+
+            Rp = AR @ self.R_lmn
+            Zp = AZ @ self.Z_lmn
+            Lp = AL @ self.L_lmn
+            surface.R_lmn = Rp
+            surface.Z_lmn = Zp
+            surface.L_lmn = Lp
             return surface
 
     def get_profile(self, name, grid=None, kind="spline", **kwargs):
@@ -1528,6 +1603,25 @@ class Equilibrium(IOAble, Optimizable):
         self._axis = new
 
     @property
+    def xsection(self):
+        """Poincare Section: Geometric surface defining toroidal cross-section."""
+        return self._xsection
+
+    @xsection.setter
+    def xsection(self, new):
+        assert isinstance(
+            new, ZernikeRZToroidalSection
+        ), f"surfaces should be of type ZernikeRZToroidalSection, got {new}"
+        assert (
+            self.sym == new.sym
+        ), "Surface and Equilibrium must have the same symmetry"
+        assert self.NFP == getattr(
+            new, "NFP", self.NFP
+        ), "Surface and Equilibrium must have the same NFP"
+        new.change_resolution(self.L, self.M, self.N)
+        self._xsection = new
+
+    @property
     def spectral_indexing(self):
         """str: Type of indexing used for the spectral basis."""
         return self._spectral_indexing
@@ -1536,11 +1630,6 @@ class Equilibrium(IOAble, Optimizable):
     def sym(self):
         """bool: Whether this equilibrium is stellarator symmetric."""
         return self._sym
-
-    @property
-    def bdry_mode(self):
-        """str: Method for specifying boundary condition."""
-        return self._bdry_mode
 
     @optimizable_parameter
     @property
@@ -1642,6 +1731,39 @@ class Equilibrium(IOAble, Optimizable):
     @Zb_lmn.setter
     def Zb_lmn(self, Zb_lmn):
         self.surface.Z_lmn = Zb_lmn
+
+    # This will become optimizable parameter automatically if user adds
+    # one of the FixSection... constraints
+    @property
+    def Rp_lmn(self):
+        """ndarray: Spectral coefficients of R at the cross-section."""
+        return self.xsection.R_lmn
+
+    @Rp_lmn.setter
+    def Rp_lmn(self, Rp_lmn):
+        self.xsection.R_lmn = Rp_lmn
+
+    # This will become optimizable parameter automatically if user adds
+    # one of the FixSection... constraints
+    @property
+    def Zp_lmn(self):
+        """ndarray: Spectral coefficients of Z at the cross-section."""
+        return self.xsection.Z_lmn
+
+    @Zp_lmn.setter
+    def Zp_lmn(self, Zp_lmn):
+        self.xsection.Z_lmn = Zp_lmn
+
+    # This will become optimizable parameter automatically if user adds
+    # one of the FixSection... constraints
+    @property
+    def Lp_lmn(self):
+        """ndarray: Spectral coefficients of Lambda at the cross-section."""
+        return self.xsection.L_lmn
+
+    @Lp_lmn.setter
+    def Lp_lmn(self, Lp_lmn):
+        self.xsection.L_lmn = Lp_lmn
 
     @optimizable_parameter
     @property
@@ -2106,6 +2228,7 @@ class Equilibrium(IOAble, Optimizable):
                     NFP=na_eq.nfp,
                 ),
                 "surface": None,
+                "xsection": None,
             }
         except AttributeError as e:
             raise ValueError("Input must be a pyQSC or pyQIC solution.") from e
@@ -2160,6 +2283,7 @@ class Equilibrium(IOAble, Optimizable):
 
         eq = Equilibrium(**inputs)
         eq.surface = eq.get_surface_at(rho=1)
+        eq.xsection = eq.get_surface_at(zeta=0)
 
         return eq
 
@@ -2206,7 +2330,6 @@ class Equilibrium(IOAble, Optimizable):
             "maxiter",
             "objective",
             "optimizer",
-            "bdry_mode",
             "output_path",
             "verbose",
         ]
@@ -2308,12 +2431,6 @@ class Equilibrium(IOAble, Optimizable):
             + "This is not recommended and may result in poor convergence. "
             + "Set grid resolutions to be higher, (i.e. eq.N_grid=2*eq.N) "
             + "to avoid this warning.",
-        )
-        errorif(
-            self.bdry_mode == "poincare",
-            NotImplementedError,
-            "Solving equilibrium with poincare XS as BC is not supported yet "
-            + "on master branch.",
         )
 
         things, result = optimizer.optimize(
@@ -2540,6 +2657,7 @@ class Equilibrium(IOAble, Optimizable):
         scales = super()._get_ess_scale(alpha, order, min_value)
         bdry_scale = self.surface._get_ess_scale(alpha, order, min_value)
         axis_scale = self.axis._get_ess_scale(alpha, order, min_value)
+        xsec_scale = self.xsection._get_ess_scale(alpha, order, min_value)
         # we use ESS for the following:
         modes = {
             "R_lmn": self.R_basis.modes,
@@ -2553,8 +2671,16 @@ class Equilibrium(IOAble, Optimizable):
 
         scales["Ra_n"] = axis_scale["R_n"]
         scales["Za_n"] = axis_scale["Z_n"]
-        scales["Rb_lmn"] = bdry_scale["R_lmn"]
-        scales["Zb_lmn"] = bdry_scale["Z_lmn"]
+        if "Rb_lmn" in self.optimizable_params:
+            scales["Rb_lmn"] = bdry_scale["R_lmn"]
+        if "Zb_lmn" in self.optimizable_params:
+            scales["Zb_lmn"] = bdry_scale["Z_lmn"]
+        if "Rp_lmn" in self.optimizable_params:
+            scales["Rp_lmn"] = xsec_scale["R_lmn"]
+        if "Zp_lmn" in self.optimizable_params:
+            scales["Zp_lmn"] = xsec_scale["Z_lmn"]
+        if "Lp_lmn" in self.optimizable_params:
+            scales["Lp_lmn"] = xsec_scale["L_lmn"]
         if hasattr(self.surface, "Phi_mn"):
             scales["Phi_mn"] = bdry_scale["Phi_mn"]
         return scales